Update to V1.10.2
Add .gitignore file to exclude the YOURLS directory, improving the process of updating the PO file from the source. This change is particularly beneficial as the YOURLS repository can be linked via a symbolic link. Additionally, update the zh_CN.po and zh_CN.mo files with new and revised translations for YOURLS version 1.10.2, incorporating new strings and adjustments to existing translations.